+++ MoltenVK/MoltenVK/Commands/MVKCommandEncoderState.h
@@ -22,6 +22,7 @@
 #include "MVKCommandResourceFactory.h"
 #include "MVKDevice.h"
 #include "MVKVector.h"
+#include <unordered_map>
 
 class MVKCommandEncoder;
 class MVKOcclusionQueryPool;
@@ -562,10 +563,13 @@ public:
 protected:
     void encodeImpl(uint32_t) override;
     void resetImpl() override;
+	bool validateOcclusionQuery(MVKOcclusionQueryPool* pQueryPool, uint32_t query);
+	void endCurrentOcclusionQuery();
 
     id<MTLBuffer> _visibilityResultMTLBuffer = nil;
     MTLVisibilityResultMode _mtlVisibilityResultMode = MTLVisibilityResultModeDisabled;
     NSUInteger _mtlVisibilityResultOffset = 0;
+	std::unordered_map<MVKQueryKey, id<MTLRenderCommandEncoder>> _mtlEncodersUsed;
 };