+++ MoltenVK/MoltenVK/GPUObjects/MVKImage.h
@@ -19,6 +19,7 @@
 #pragma once
 
 #include "MVKResource.h"
+#include "MVKCommandResourceFactory.h"
 #include "MVKSync.h"
 #include "MVKVector.h"
 #include <MoltenVKSPIRVToMSLConverter/SPIRVToMSLConverter.h>
@@ -30,8 +31,6 @@
 class MVKImageView;
 class MVKSwapchain;
 class MVKCommandEncoder;
-struct MVKImageDescriptorData_t;
-typedef MVKImageDescriptorData_t MVKImageDescriptorData;
 
 
 /** Tracks the state of an image subresource.  */
@@ -398,16 +397,6 @@ protected:
 
 #pragma mark -
 #pragma mark MVKSwapchainImage
-
-/** Indicates the relative availability of each image in the swapchain. */
-typedef struct MVKSwapchainImageAvailability_t {
-	uint64_t acquisitionID;			/**< When this image was last made available, relative to the other images in the swapchain. Smaller value is earlier. */
-	uint32_t waitCount;				/**< The number of semaphores already waiting for this image. */
-	bool isAvailable;				/**< Indicates whether this image is currently available. */
-
-	bool operator< (const MVKSwapchainImageAvailability_t& rhs) const;
-} MVKSwapchainImageAvailability;
-
 
 /** Represents a Vulkan image used as a rendering destination within a swapchain. */
 class MVKSwapchainImage : public MVKImage {